pyproject.toml
[tool.poetry]
name = "church-web-helper"
version = "1.8.0"
description = "A python package to make use of ChurchTools API and Communi API with a docker packaged WebUI"
authors = [
    "bensteUEM",
]
homepage = "https://github.com/bensteUEM/ChurchWebHelper"
license = "CC-BY-SA"
readme = "README.md"
include = [
    "templates/*.html",
    "static/*",
]

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^2.3.2"
Flask-Session = "^0.5.0"
requests = "^2.31.0"
python-docx = "^0.8.11"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.0"
matplotlib = "^3.9.1"
pandas = "^2.2.2"
toml = "^0.10.2"
vobject = "^0.9.8"
xlsxwriter = "^3.2.0"
tzlocal = "^5.2"

[tool.poetry.dependencies.churchtools-api]
git = "https://github.com/bensteUEM/ChurchToolsAPI.git"
branch = "main"

[tool.poetry.dependencies.communi-api]
git = "https://github.com/bensteUEM/CommuniAPI.git"
branch = "main"

[tool.poetry.group.dev.dependencies]
poetry = "^1.6.1"
tomli_w = "^1.0.0"
wheel = "^0.41.2"
setuptools = "^66.1.1"
autopep8 = "^2.0.4"
pytest = "^8.3.4"
ruff = "^0.9.3"

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 88
indent-width = 4
target-version = "py310"
output-format = "grouped"

[tool.ruff.lint]
select = [
    "ALL",
]
ignore = [
    "FIX002",
    "COM812",
    "ISC001",
]
fixable = [
    "ALL",
]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = [
    "S101",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[build-system]
requires = [
    "poetry-core",
]
build-backend = "poetry.core.masonry.api"
